12路多路高压箱导出逻辑修改为值为负数,删除不需要的日志

+++ UtilityTools/Modules/UtilityTools.Modules.FDCCHVBox/Model/HVBoxModel.cs
@@ -627,7 +627,7 @@ namespace UtilityTools.Modules.FDC12CHVBox.Model
         public DateTime  DateTime { get; set; }
         public float HV { get; set; }
         public float I { get; set; }
-        public ushort setHv { get; set; }
+        public int setHv { get; set; }
     }
   
 }
